+++ crates/scheduler/src/event/button.rs
@@ -17,6 +17,7 @@ use alloc::vec::Vec;
 use derive_where::derive_where;
 use wasefire_board_api::button::Event;
 use wasefire_board_api::{self as board, Api as Board, Id};
+use wasefire_error::Error;
 
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
 #[derive_where(Debug, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
@@ -33,6 +34,13 @@ impl<B: Board> From<Key<B>> for crate::event::Key<B> {
 impl<'a, B: Board> From<&'a Event<B>> for Key<B> {
     fn from(event: &'a Event<B>) -> Self {
         Key { button: event.button }
+    }
+}
+
+impl<B: Board> Key<B> {
+    pub fn disable(self) -> Result<(), Error> {
+        use wasefire_board_api::button::Api as _;
+        board::Button::<B>::disable(self.button)
     }
 }