+++ crates/scheduler/src/protocol.rs
@@ -21,15 +21,14 @@ use wasefire_board_api::{self as board, Api as Board};
 use wasefire_error::{Code, Error};
 use wasefire_logger as log;
 use wasefire_protocol::applet::AppletId;
+#[cfg(feature = "wasm")]
+use wasefire_protocol::applet::ExitStatus;
 use wasefire_protocol::{self as service, Api, ApiResult, Request, Service, VERSION};
 
 use crate::{Applet, Scheduler};
 
 #[derive(Debug, Default)]
-pub struct State {
-    tunnel: TunnelState,
-    update: TransferState,
-}
+pub struct State(StateImpl);
 
 pub fn enable<B: Board>() {
     if let Err(error) = board::platform::Protocol::<B>::enable() {
@@ -56,11 +55,12 @@ pub fn process_event<B: Board>(scheduler: &mut Scheduler<B>, event: board::Event
 fn process_event_<B: Board>(
     scheduler: &mut Scheduler<B>, event: board::Event<B>, request: Box<[u8]>,
 ) -> Result<(), Error> {
-    match &scheduler.protocol.tunnel {
-        TunnelState::Normal => (),
-        TunnelState::Tunnel { applet_id, delimiter } => {
+    match &scheduler.protocol.0 {
+        Normal { .. } => (),
+        Locked => return Err(Error::user(Code::InvalidState)),
+        Tunnel { applet_id, delimiter } => {
             if request == *delimiter {
-                scheduler.protocol.tunnel = TunnelState::Normal;
+                scheduler.protocol = State::default();
                 return Ok(reply::<B, service::AppletTunnel>(()));
             }
             return applet::<B>(scheduler, *applet_id)?.put_request(event, &request);
@@ -82,14 +82,9 @@ fn process_event_<B: Board>(
             board::Platform::<B>::reboot()?;
         }
         Api::AppletTunnel(service::applet::Tunnel { applet_id, delimiter }) => {
-            match scheduler.protocol.tunnel {
-                TunnelState::Normal => {
-                    let delimiter = delimiter.to_vec().into_boxed_slice();
-                    scheduler.protocol.tunnel = TunnelState::Tunnel { applet_id, delimiter };
-                    reply::<B, service::AppletTunnel>(())
-                }
-                TunnelState::Tunnel { .. } => unreachable!(),
-            }
+            let delimiter = delimiter.to_vec().into_boxed_slice();
+            scheduler.protocol.0 = Tunnel { applet_id, delimiter };
+            reply::<B, service::AppletTunnel>(())
         }
         Api::PlatformInfo(()) => {
             use wasefire_board_api::platform::Api as _;
@@ -108,7 +103,36 @@ fn process_event_<B: Board>(
             let response = board::platform::Update::<B>::metadata()?;
             reply::<B, service::PlatformUpdateMetadata>(&response);
         }
-        Api::PlatformUpdateTransfer(request) => process_transfer::<B>(scheduler, request)?,
+        Api::PlatformUpdateTransfer(request) => process_update::<B>(scheduler, request)?,
+        #[cfg(feature = "native")]
+        Api::AppletInstall(_) | Api::AppletUninstall(_) => {
+            return Err(Error::world(Code::NotImplemented));
+        }
+        #[cfg(feature = "wasm")]
+        Api::AppletInstall(request) => process_install::<B>(scheduler, request)?,
+        #[cfg(feature = "wasm")]
+        Api::AppletUninstall(()) => {
+            use wasefire_board_api::applet::Api as _;
+            board::Applet::<B>::start(false)?;
+            board::Applet::<B>::finish()?;
+            scheduler.stop_applet(ExitStatus::Kill);
+            reply::<B, service::AppletUninstall>(());
+        }
+        Api::AppletExitStatus(applet_id) => {
+            use crate::applet::Slot;
+            let service::applet::AppletId = applet_id;
+            let status = match scheduler.applet {
+                #[cfg(feature = "wasm")]
+                Slot::Empty => return Err(Error::user(Code::NotFound)),
+                Slot::Running(_) => None,
+                Slot::Exited(x) => Some(x),
+            };
+            reply::<B, service::AppletExitStatus>(status);
+        }
+        Api::PlatformLock(()) => {
+            scheduler.protocol.0 = Locked;
+            reply::<B, service::PlatformLock>(());
+        }
         #[cfg(not(feature = "_test"))]
         _ => return Err(Error::internal(Code::NotImplemented)),
     }
@@ -128,9 +152,10 @@ pub fn put_response<B: Board>(
         }
         Err(error) => return Err(error),
     }
-    match &call.scheduler().protocol.tunnel {
-        TunnelState::Normal => Ok(()),
-        TunnelState::Tunnel { applet_id, .. } => {
+    match &call.scheduler().protocol.0 {
+        Normal { .. } => Ok(()),
+        Locked => Err(Error::world(Code::InvalidState)),
+        Tunnel { applet_id, .. } => {
             let service::applet::AppletId = applet_id;
             match call.applet().get_response()? {
                 Some(response) => board::platform::Protocol::<B>::write(&response),
@@ -143,46 +168,63 @@ pub fn put_response<B: Board>(
     }
 }
 
-fn process_transfer<B: Board>(
+fn process_update<B: Board>(
     scheduler: &mut Scheduler<B>, request: service::transfer::Request,
 ) -> Result<(), Error> {
     use wasefire_board_api::platform::update::Api as _;
     match request {
         service::transfer::Request::Start { dry_run } => {
-            scheduler.protocol.update = TransferState::Running { dry_run };
+            *scheduler.protocol.0.update() = TransferState::Running { dry_run };
             board::platform::Update::<B>::initialize(dry_run)?;
         }
         service::transfer::Request::Write { chunk } => {
-            let _ = scheduler.protocol.update.dry_run()?;
+            let _ = scheduler.protocol.0.update().dry_run()?;
             board::platform::Update::<B>::process(chunk)?;
         }
         service::transfer::Request::Finish => {
-            let _ = scheduler.protocol.update.dry_run()?;
+            let _ = scheduler.protocol.0.update().dry_run()?;
             board::platform::Update::<B>::finalize()?;
-            scheduler.protocol.update = TransferState::Ready;
+            *scheduler.protocol.0.update() = TransferState::Ready;
         }
     }
     reply::<B, service::PlatformUpdateTransfer>(());
     Ok(())
 }
 
-#[derive(Debug, Default)]
-enum TunnelState {
-    #[default]
-    Normal,
-    Tunnel {
-        applet_id: service::applet::AppletId,
-        delimiter: Box<[u8]>,
-    },
+#[cfg(feature = "wasm")]
+fn process_install<B: Board>(
+    scheduler: &mut Scheduler<B>, request: service::transfer::Request,
+) -> Result<(), Error> {
+    use wasefire_board_api::applet::Api as _;
+    match request {
+        service::transfer::Request::Start { dry_run } => {
+            if !dry_run {
+                scheduler.stop_applet(ExitStatus::Kill);
+            }
+            *scheduler.protocol.0.install() = TransferState::Running { dry_run };
+            board::Applet::<B>::start(dry_run)?;
+        }
+        service::transfer::Request::Write { chunk } => {
+            let _ = scheduler.protocol.0.install().dry_run()?;
+            board::Applet::<B>::write(chunk)?;
+        }
+        service::transfer::Request::Finish => {
+            let dry_run = scheduler.protocol.0.install().dry_run()?;
+            board::Applet::<B>::finish()?;
+            *scheduler.protocol.0.install() = TransferState::Ready;
+            if !dry_run {
+                scheduler.start_applet()?;
+            }
+        }
+    }
+    reply::<B, service::AppletInstall>(());
+    Ok(())
 }
 
-#[derive(Debug, Default)]
+#[derive(Debug)]
 enum TransferState {
-    #[default]
     Ready,
-    Running {
-        dry_run: bool,
-    },
+    Running { dry_run: bool },
 }
 
 impl TransferState {
@@ -194,11 +236,53 @@ impl TransferState {
     }
 }
 
+#[derive(Debug)]
+enum StateImpl {
+    Normal {
+        update: TransferState,
+        #[cfg(feature = "wasm")]
+        install: TransferState,
+    },
+    Locked,
+    Tunnel {
+        applet_id: service::applet::AppletId,
+        delimiter: Box<[u8]>,
+    },
+}
+use StateImpl::*;
+
+impl Default for StateImpl {
+    fn default() -> Self {
+        Normal {
+            update: TransferState::Ready,
+            #[cfg(feature = "wasm")]
+            install: TransferState::Ready,
+        }
+    }
+}
+
+impl StateImpl {
+    fn update(&mut self) -> &mut TransferState {
+        match self {
+            Normal { update, .. } => update,
+            _ => unreachable!(),
+        }
+    }
+
+    #[cfg(feature = "wasm")]
+    fn install(&mut self) -> &mut TransferState {
+        match self {
+            Normal { install, .. } => install,
+            _ => unreachable!(),
+        }
+    }
+}
+
 fn applet<B: Board>(
     scheduler: &mut Scheduler<B>, applet_id: AppletId,
 ) -> Result<&mut Applet<B>, Error> {
     let AppletId = applet_id;
-    scheduler.applet.as_mut().ok_or_else(|| {
+    scheduler.applet.get().ok_or_else(|| {
         log::warn!("Failed to find applet");
         Error::world(Code::NotFound)
     })