+++ crates/cli/CHANGELOG.md
@@ -9,10 +9,14 @@
 
 ### Minor
 
+- Add `platform-lock` to lock a platform protocol
+- Add `applet-exit-status` to get an applet exit status
+- Implement `applet-{install,uninstall}` for applet management
 - Add `platform-update-{metadata,transfer}` for platform update
 
 ### Patch
 
+- Print `wasefire` instead of `wasefire-cli` with `--version`
 - Check for bad command-line configuration
 - Update dependencies
 - Restore release builds to the default
@@ -40,4 +44,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 7 -->
+<!-- Increment to skip CHANGELOG.md test: 8 -->